pyproject.toml
[project]
name = "secrin"
version = "0.1.0"
description = ""
authors = [{ name = "Jenil Savani", email = "jenil@secrinlabs.com" }]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.120.1,<0.121.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "neo4j (>=6.0.2,<7.0.0)",
    "tree-sitter (>=0.25.2,<0.26.0)",
    "gitpython (>=3.1.45,<4.0.0)",
    "tree-sitter-python (>=0.25.0,<0.26.0)",
    "tree-sitter-javascript (>=0.25.0,<0.26.0)",
]

[tool.poetry]
packages = [{ include = "packages" }]

[tool.poetry.scripts]
migrate = "packages.database.graph.migrations.run_migrations:run_migrations"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
